requirements.txt
Flask>=2.1
geopy>=2.3
opencv-python>=4.7
werkzeug>=2.1
python-dotenv>=0.21.0
Flask-SQLAlchemy>=3.0
pytest>=7.0
alembic>=1.10
gunicorn>=20.1

# Optional (only if you use additional libs)
# pytest>=7.0     # for tests
Flask>=2.1
geopy>=2.3
opencv-python>=4.7
werkzeug>=2.1
python-dotenv>=0.21.0
Flask-SQLAlchemy>=3.0
pytest>=7.0
alembic>=1.10

# Optional (only if you use additional libs)
# gunicorn>=20.1  # for production WSGI server
# pytest>=7.0     # for tests
